+++ src/tests/membench.c
@@ -14,8 +14,8 @@
 #define MAX_ALLOC_MULTIPLIER 4
 
 typedef struct {
-	size_t num_allocations;
-	size_t alloc_size;
+	uint64_t num_allocations;
+	uint64_t alloc_size;
 	init_allocator_fn my_init;
 	destroy_allocator_fn my_destroy;
 	malloc_fn my_malloc;
@@ -48,15 +48,15 @@ static void* thread_fn(void* arg) {
 	// -----------------------------------
 
 	// Make N allocations of random size
-	for(size_t i = 0; i < ctx->num_allocations; ++i) {
-		const size_t size = ctx->alloc_size * (1 + rand_r(&seed) % MAX_ALLOC_MULTIPLIER);
+	for(uint64_t i = 0; i < ctx->num_allocations; ++i) {
+		const uint64_t size = ctx->alloc_size * (1 + rand_r(&seed) % MAX_ALLOC_MULTIPLIER);
 		ptrs[i] = ctx->my_malloc(size);
 		assert(ptrs[i] != NULL);
 		memset(ptrs[i], 0xFF, size);
 	}
 
 	// Free ~50% of allocations
-	for(size_t i = 0; i < ctx->num_allocations; ++i) {
+	for(uint64_t i = 0; i < ctx->num_allocations; ++i) {
 		if(rand_r(&seed) % 2 == 0) {
 			ctx->my_free(ptrs[i]);
 			ptrs[i] = NULL;
@@ -64,16 +64,16 @@ static void* thread_fn(void* arg) {
 	}
 
 	// Re-allocate those ~50%
-	for(size_t i = 0; i < ctx->num_allocations; ++i) {
+	for(uint64_t i = 0; i < ctx->num_allocations; ++i) {
 		if(ptrs[i] == NULL) {
-			const size_t size = ctx->alloc_size * (1 + rand_r(&seed) % MAX_ALLOC_MULTIPLIER);
+			const uint64_t size = ctx->alloc_size * (1 + rand_r(&seed) % MAX_ALLOC_MULTIPLIER);
 			ptrs[i] = ctx->my_malloc(size);
 			assert(ptrs[i] != NULL);
 		}
 	}
 
 	// Free all allocations
-	for(size_t i = 0; i < ctx->num_allocations; ++i) {
+	for(uint64_t i = 0; i < ctx->num_allocations; ++i) {
 		ctx->my_free(ptrs[i]);
 	}
 
@@ -92,14 +92,14 @@ static void* thread_fn(void* arg) {
 	return (void*)(intptr_t)(after - before);
 }
 
-static double run_config(size_t num_threads, thread_context* ctx) {
+static double run_config(uint64_t num_threads, thread_context* ctx) {
 	pthread_t thread_ids[num_threads];
-	for(size_t i = 0; i < num_threads; ++i) {
+	for(uint64_t i = 0; i < num_threads; ++i) {
 		pthread_create(&thread_ids[i], NULL, thread_fn, ctx);
 	}
 
 	double time_sum = 0.0;
-	for(size_t i = 0; i < num_threads; ++i) {
+	for(uint64_t i = 0; i < num_threads; ++i) {
 		intptr_t delta_time;
 		pthread_join(thread_ids[i], (void**)&delta_time);
 		time_sum += delta_time;
@@ -114,15 +114,15 @@ static void run_membench(init_allocator_fn my_init, destroy_allocator_fn my_dest
 		my_init(POOL_SIZE);
 	}
 
-	const size_t configs[][3] = {
+	const uint64_t configs[][3] = {
 		{ 1, 1000, 256 }, { 10, 1000, 256 }, { 50, 1000, 256 }, { 100, 1000, 32 }
 	};
-	const size_t num_configs = sizeof(configs) / sizeof(size_t[3]);
+	const uint64_t num_configs = sizeof(configs) / sizeof(uint64_t[3]);
 
-	for(size_t i = 0; i < num_configs; ++i) {
-		const size_t num_threads = configs[i][0];
-		const size_t num_allocations = configs[i][1];
-		const size_t alloc_size = configs[i][2];
+	for(uint64_t i = 0; i < num_configs; ++i) {
+		const uint64_t num_threads = configs[i][0];
+		const uint64_t num_allocations = configs[i][1];
+		const uint64_t alloc_size = configs[i][2];
 		thread_context system_ctx = { .num_allocations = num_allocations,
 			                          .alloc_size = alloc_size,
 			                          .my_init = NULL,
